requirements.txt
python-jobspy
discord.py
pandas
python-dotenv
schedule
